+++ src/odbc_result.cpp
@@ -286,7 +286,9 @@ nanodbc::timestamp odbc_result::as_timestamp(double value) {
   auto utc_time = system_clock::from_time_t(static_cast<std::time_t>(value));
 
   auto civil_time = cctz::convert(utc_time, c_->timezone());
-  ts.fract = frac;
+  // We are using a fixed precision of 3, as that is all we can be guaranteed
+  // to support in SQLServer
+  ts.fract = (std::int32_t)(frac * 1000) * 1000000;
   ts.sec = civil_time.second();
   ts.min = civil_time.minute();
   ts.hour = civil_time.hour();