+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Features
 
+* Subseconds are now retained when inserting POSIXct objects (#130, #208)
+
 * `dbSendStatement()` now supports the `params` argument, which allows it to be
   used (indirectly) by `DBI::dbAppendTable()` (#215).